requirements.txt
gradio
torch==2.0.0+cu118
torchvision==0.15.0+cu118
torchaudio==2.0.1+cu118
--extra-index-url https://download.pytorch.org/whl/cu118
transformers
pillow
numpy
einops
pulp
scipy
shapely
timm